refactor(router): remove history property
BREAKING CHANGE: the history property was marked as internal already. Since we
need to pass the history instance to the router, we always have access to it,
differently from Vue Router 3 where the history was instantiated internally.
The history API was also internal (it wasn't documented), so this change
shouldn't be a problem as people shouldn't be relying on `router.history` in
their apps. If you think this property is needed, please open an issue to
discuss the use case. Note it's already accessible as you have to create it:
```js
export const history = createWebHistory()
export const router = createRouter({ history, routes: [] })
```